Add *mysql-encoding* to enable forcing the encoding.
It so happens that MySQL stores data in an encoding that is different
from what it knows in the meta-data, and that the application knows the
real data encoding. With that dynamic binding it's now possible to force
the Qmynd driver using the known data encoding, which might very well be
different from the meta-data.