Fix multiple packet protocol implementation.
As seen in the following pgloader issue Qmynd implementation of multi
packet MySQL protocol was incomplete and faulty. This patch fixes it to
actually fetch the next packet transparently when needed, which was
always the intend of the code.

In passing, introduce support for the max_allowed_packets idea that is
found in the MySQL configuration, server side, and available in the
default mysql client.

References:

  https://github.com/dimitri/pgloader/issues/328
  https://dev.mysql.com/doc/internals/en/sending-more-than-16mbyte.html