Desultory pass at improving efficiency of the driver.
Handling the different choice combinations within the hot-spot function is
not a very good idea. Instead define four specialized functions and arrange
to pick which to call from upper in the call stack.

Also replace a series of member calls to a simple case form, which is much
simpler to execute.

Good performances results are seen from pgloader side of things. Nothing
spectacular, but every bit counts I guess.