Fix PROTOCOL-VERSION-MISMATCH (255).
As we can see in the details of the bug investigation in
https://github.com/dimitri/pgloader/issues/657, the MySQL protocol may
return an error packet as an answer to opening a connection, before the
initial handshake.

The error packet has the tag 0xff or 255, and Qmynd used to parse it as a
protocol version number. This patch fixes it and introduces a new packet
type 'response-error-no-sql-state', because at this point in the
communication no capabilities are negociated yet, so the error packet does
not contain the SQL state.

Fixes https://github.com/dimitri/pgloader/issues/657.