Fix previous patch and improve error reporting.
Previous patch didn't consider that case doesn't evaluate its branch labels,
so we have to #. (list ...) our constants here. What happened with the
previous coding was that the 'otherwise case would be used always, so that
we only have errors when using BINARY data, basically.

See https://github.com/dimitri/pgloader/issues/716 for such an error case.

In passing, improve the error handling of babel decoding errors by reporting
our own error with information about the MySQL column in which we failed to
read and decode the data. That's pretty useful actually.